+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Improvements:
 * Locale: the missing items are now sorted by application, type, module and name.
 * Core: don't add a timestamp to the urls of well known libraries in the backend.
 * Core: automagic canonical-urls.
+* Core: added a new modifier stripnewlines which will remove all newlines in a string, so JS can handle it. 
 
 Bugfixes:
 * Pages: default blocks now apply correctly on new pages.